--- R3_VillagePeople_Mahtimokit/Order_tab.cs
@@ -1,81 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace R3_VillagePeople_Mahtimokit
-{
-    public partial class cal_Order_Cottage_end : Form
-    {
-        public cal_Order_Cottage_end()
-        {
-            InitializeComponent();
-        }
-
-        private void Varaus_uusi_Load(object sender, EventArgs e)
-        {
-
-        }
-
-        private void comboBox2_SelectedIndexChanged(object sender, EventArgs e)
-        {
-
-        }
-
-        private void button13_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label13_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label8_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label15_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label16_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label14_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label17_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void groupBox1_Enter(object sender, EventArgs e)
-        {
-
-        }
-
-        private void toimipisteetToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            Management_tab lomake = new Management_tab();
-            lomake.Show();
-        }
-
-        private void listView8_SelectedIndexChanged(object sender, EventArgs e)
-        {
-
-        }
-    }
-}
